+++ scripts/update.js
@@ -12,7 +12,8 @@ export async function checkUpdate() {
             .replaceAll('New Content:', '<b><u>New Content:</u></b>')
             .replaceAll('New Monster Content:', '<b><u>New Monster Content:</u></b>')
             .replaceAll('Bug Fixes:', '<b><u>Bug Fixes:</u></b>')
-            .replaceAll('Update Notes:', '<b><u>Update Notes:</u></b>');
+            .replaceAll('Update Notes:', '<b><u>Update Notes:</u></b>')
+            .replaceAll('New Enhanced Animations:', '<b><u>New Enhanced Animations:</u></b>');
         let message = '<hr>Chris\'s Premades update <b>' + info.tag_name + '</b> available!<hr>' + body;
         await ChatMessage.create({
             'speaker': {'alias': name},